requirements.txt
pyautogui >= 0.9.54
opencv-python >= 4.10
pyscreeze >= 0.1.30
pillow >= 10.4.0
dbus-python >= 1.3.2
pyscreenshot >= 3.1
